Split ResetPassword - created dedicated form, updated success & error variables
